backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
ldap3==2.9.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
email-validator==2.2.0
pywin32==311
pywinrm==0.4.3
